Lab3 Adv1 complete

+++ lab3/include/priority_queue.h
@@ -1,8 +1,8 @@
 #pragma once
-int add_node(void (*callback_f)(),void* arguments,float times, float time_gap);
+int add_node(void (*callback_f)(),void* arguments,int times, int time_gap);
 struct node* delete_first_node();
 struct node{
-    float time_to_ring;
+    int time_to_ring;
     void (*todo)();
     void *arguments;
     struct node* next;
